+++ src/compas/datastructures/halfedge/__init__.py
@@ -1,0 +1,1 @@
+from .halfedge import HalfEdge  # noqa: F401