+++ CHANGELOG.md
@@ -9,6 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+
+## [1.8.1] 2021-09-08
+
+### Added
+
+### Changed
+
+### Removed
+
+
+## [1.8.0] 2021-09-08
+
+### Added
+
 * Added pluggable function `trimesh_slice` in `compas_rhino`.
 * Added equality comparison for pointclouds.
 * Added `compas.data.is_sequence_of_uint`.
@@ -16,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added support for multimesh files to OBJ reader/writer.
 * Added support for attaching and detaching meshes in `compas.robots.RobotModelArtist` and drawing them.
 * Added `meshes` method to artists of `compas.robots.RobotModel`.
+* Added `FrameArtist` class to `compas_blender`.
 
 ### Changed
 
@@ -27,8 +46,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed `compas_rhino.drawing.draw_breps` to assume provided polygon is closed and automatically add missing corner to polycurve constructor.
 * Changed conversion of edges and faces to uniques keys for the data dicts to use the string representation of a sorted tuple of identifiers.
 * Added `dtype` to JSON decoding error message.
+* Moved `compas.datastructures.mesh.core.halfedge.HalfEdge` to `compas.datastructures.halfedge.halfedge.HalfEdge`
+* Moved `compas.datastructures.network.core.graph.Graph` to `compas.datastructures.graph.graph.Graph`.
 
 ### Removed
+
+* Removed `compas.datastructures.mesh.core.mesh.BaseMesh`.
+
+* Removed `compas.datastructures.BaseNetwork`.
 
 ## [1.7.1] 2021-06-14